Rename IChartState to IPackageState
