+++ dashboard/src/components/AppUpgrade/AppUpgrade.tsx
@@ -11,7 +11,7 @@ import { useDispatch, useSelector } from "react-redux";
 import * as ReactRouter from "react-router";
 import { Action } from "redux";
 import { ThunkDispatch } from "redux-thunk";
-import { FetchError, IAppRepository, IChartState, IStoreState, UpgradeError } from "shared/types";
+import { FetchError, IAppRepository, IPackageState, IStoreState, UpgradeError } from "shared/types";
 import LoadingWrapper from "../LoadingWrapper/LoadingWrapper";
 import SelectRepoForm from "../SelectRepoForm/SelectRepoForm";
 import UpgradeForm from "../UpgradeForm/UpgradeForm";
@@ -26,8 +26,8 @@ export interface IAppUpgradeProps {
   releaseName: string;
   repoName?: string;
   repoNamespace?: string;
-  selected: IChartState["selected"];
-  deployed: IChartState["deployed"];
+  selected: IPackageState["selected"];
+  deployed: IPackageState["deployed"];
   reposIsFetching: boolean;
   repoError?: Error;
   chartsError: Error | undefined;
