+++ dashboard/src/components/DeploymentFormBody/DeploymentFormBody.tsx
@@ -5,7 +5,7 @@ import Tabs from "components/Tabs";
 import { isEqual } from "lodash";
 import { useEffect, useState } from "react";
 import { parseValues, retrieveBasicFormParams, setValue } from "../../shared/schema";
-import { DeploymentEvent, IBasicFormParam, IChartState } from "../../shared/types";
+import { DeploymentEvent, IBasicFormParam, IPackageState } from "../../shared/types";
 import { getValueFromEvent } from "../../shared/utils";
 import ConfirmDialog from "../ConfirmDialog/ConfirmDialog";
 import LoadingWrapper from "../LoadingWrapper/LoadingWrapper";
@@ -20,7 +20,7 @@ export interface IDeploymentFormBodyProps {
   chartVersion: string;
   deployedValues?: string;
   chartsIsFetching: boolean;
-  selected: IChartState["selected"];
+  selected: IPackageState["selected"];
   appValues: string;
   setValues: (values: string) => void;
   setValuesModified: () => void;
